Add an AppCommon framework
