+++ app/Navigation.swift
@@ -1,0 +1,12 @@
+//
+//  Navigation.swift
+//  AccountsJourney
+//
+//  Created by George Nyakundi on 03/02/2025.
+//
+
+import Foundation
+
+enum Navigation {
+    private static var defaultNavigation: ()
+}
