+++ appCommon/Dependencies/UserInactivityTracker.swift
@@ -1,6 +1,6 @@
 //
 //  UserInactivityTracker.swift
-//  AccountsJourney
+//  AppCommon
 //
 //  Created by George Nyakundi on 03/02/2025.
 //
