+++ appCommon/Extensions/AppDelegate+MSDK.swift
